+++ front/lib/swr/tags.ts
@@ -232,33 +232,3 @@ export function useUpdateAgentTags({ owner }: { owner: LightWorkspaceType }) {
 
   return updateAgentTags;
 }
-
-export function useBatchUpdateAgentTags({
-  owner,
-}: {
-  owner: LightWorkspaceType;
-}) {
-  const batchUpdateAgentTags = useCallback(
-    async (
-      agentIds: string[],
-      body: { addTagIds?: string[]; removeTagIds?: string[] }
-    ) => {
-      await fetch(
-        `/api/w/${owner.sId}/assistant/agent_configurations/batch_update`,
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({
-            agentIds,
-            ...body,
-          }),
-        }
-      );
-    },
-    [owner]
-  );
-
-  return batchUpdateAgentTags;
-}
